pyproject.toml
# syntax-documentation:
#  - https://python-poetry.org/docs/pyproject
#  - https://flit.readthedocs.io/en/latest/pyproject_toml.html
#
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.
#
#   > poetry install
#   > poetry install --only docs
#   > poetry install (--with|--without) (docs|dev|ipython|profiling|testing|typing)

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
target-version = ['py39']


[tool.poetry]
name = "scouts-kampvisum-api"
version = "1.0"
description = "Provides functionality for setting up a scouts kamp visum"
authors = ["inuits <info@inuits.eu>"]
# license = "GPL v2.0"
homepage = "https://github.com/ScoutsGidsenVL/kampvisum-backend"
repository = "https://github.com/ScoutsGidsenVL/kampvisum-backend"
keywords = ["django", "scouts", "kampvisum"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
boto3 = ">=1.34.39"
cffi = ">=1.16.0"
django = ">=4.1.13,<5.0"
django-anymail = {extras = ["sendinblue"], version = ">=10.2"}
django-cors-headers = ">=4.3.1"
django-cprofile-middleware = ">=1.0.5"
django-extensions = ">=3.2.3"
django-filter = ">=23.5"
django-rest-enumfield = ">=0.2.0"
django-safedelete = ">=1.3.3"
django-sql-middleware = ">=0.0.7"
django-storages = ">=1.14.2"
djangorestframework = ">=3.14,<3.15"
drf-yasg = ">=1.21.4"
environs = ">=10.3.0"
future = ">=0.18.3"
mozilla-django-oidc = ">=4.0.0"
pdfrw = ">=0.4"
psycopg2-binary = ">=2.9.9"
pyjwt = ">=2.8.0"
PyYAML = ">=6.0.1"
urllib3= "1.26.18"

[tool.poetry.group.dev.dependencies]
autoread-dotenv = ">=1.0.2"
pylint = ">=3.1.0"
pylint-django = ">=2.5.5"
ruff = ">=0.0"

[tool.poetry.group.ipython.dependencies]
ipdb = ">=0.13"
ipython = ">=8.18"

[tool.poetry.group.testing.dependencies]
coverage = { extras = ["toml"], version = ">=6.2" }
factory-boy = ">=3.3.0"
hypothesis = ">=6.72"
# nox = ">=2022.11"
# nox-poetry = ">=1.0"
pytest = ">=7.0"
pytest-clarity = ">=1.0"
pytest-codecov = ">=0.5"
pytest-cov = ">=3.0"
pytest-click = ">=1.1"
pytest-django = ">=4.5.2"
pytest-mock = ">=3.6"
pytest-xdist = ">=3.2"

[tool.poetry.group.typing.dependencies]
django-stubs = ">=4.2.7"
mypy = ">=1.9"

[tool.pylint.format]
max-line-length = 120
good-names = [
    # "foo", # dummy variable
    # "i",   # counter in loop
    # "ok",  # status
]

[tool.pylint.linters]
max-line-length = 120
good-names = [
    # "foo", # dummy variable
    # "i",   # counter in loop
    # "ok",  # status
]
ignored-modules = []
ignore-paths=[
    "src/.*/migrations/",
    "scouts_wwdb_api/.*/migrations/",
]

[tool.pylint.messages_control]
disable = [
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    # "E1101",  # no-member
    "R0903",  # too-few-public-methods
    "W0223",  # abstract-method
    "W1203",  # logging-fstring-interpolation
]

[tool.pylint.typecheck]
generated-members= ["save"] #REQUEST,acl_users,aq_parent,"[a-zA-Z]+_set{1,2}",save,delete


[tool.pytest.ini_options]
# cache_dir = "var/cache/pytest"  # relative to $PWD, use PYTEST_CACHE_DIR instead
log_cli = false                   # enable to show log-output
log_cli_level = "NOTSET"
filterwarnings = []
markers = ["unit", "integration"]
testpaths = ["tests"]
# the junit-report is used to report coverage in gitlab
addopts = "--color=yes --junit-xml='var/coverage/pytest.xml'"
DJANGO_SETTINGS_MODULE = "scouts_kampvisum_api.settings"
pythonpath = [".", "scouts_kampvisum_api"]


[tool.ruff]
# ruff searches for settings in the target-dir and up
# https://docs.astral.sh/ruff/configuration
cache-dir = "var/cache/ruff"
line-length = 120
